fix compiler warnings
DEBUG should be set by "configure", this was done on purpose so when you build
the CDK, the global Makefile can decide if the applications should print
additional debug messages or not.

If a developer thinks his application is flawless and no further debug output
is necessary he/she should rename the DEBUG to something else but not shadow it.